Change format to double

+++ src/main/kotlin/com/go/feature/component/filter/util/TypeConverterUtils.kt
@@ -1,8 +1,8 @@
 package com.go.feature.component.filter.util
 
-fun parseLong(field: String, value: String?): Long? {
+fun parseDouble(field: String, value: String?): Double? {
     return try {
-        value?.toLong()
+        value?.toDouble()
     } catch (e: NumberFormatException) {
         throw IllegalArgumentException("Value is not compatible with filter; field=${field}, value=${value}")
     }
